Collect_action class + associated test and some change on villager and resources_point (new get)

+++ core/resource.py
@@ -20,26 +20,26 @@ class Resource:
     def get_food(self) -> int:
         return self.__food
     
-    def add_wood(self,value : int) :
+    def add_wood(self,value : float) :
         assert value >= 0, value
         self.__wood += value
 
-    def add_gold(self,value : int) :
+    def add_gold(self,value : float) :
         assert value >= 0, value
         self.__gold += value
 
-    def add_food(self,value : int) :
+    def add_food(self,value : float) :
         assert value>=0,value
         self.__food += value
 
-    def remove_wood(self,value : int) :
+    def remove_wood(self,value : float) :
         assert self.__wood >= value >= 0,value
         self.__wood -= value
 
-    def remove_gold(self,value : int) :
+    def remove_gold(self,value : float) :
         assert self.__gold >= value >= 0,value
         self.__gold -= value
 
-    def remove_food(self,value : int) :
+    def remove_food(self,value : float) :
         assert self.__food >= value >= 0,value
         self.__food -= value
